Added select and delete operations

+++ run.sh
@@ -1,6 +1,10 @@
 #! /usr/bin/env bash
 
-rm -rf build
-cmake -B build &&
-    cmake --build build &&
+if [[ "$@" == "run" ]]; then
     ./build/cpp_db
+else
+    rm -rf build
+    cmake -B build &&
+        cmake --build build &&
+        ./build/cpp_db
+fi
